scripts/requirements.txt
BeautifulSoup==3.2.1
python-espncricinfo==0.2.2
requests==2.7.0
simplejson==3.7.3
six==1.9.0
